Break out API into domains, modify bump/increment API to act like fetch/get and be chainable.
