+++ lib/mnemonix/store/bump/functions.ex
@@ -6,28 +6,57 @@ defmodule Mnemonix.Store.Bump.Functions do
     quote location: :keep do
 
       @doc false
-      def increment(store, key, amount \\ 1), do: do_bump(store, :increment, key, amount)
+      def bump(store, key, amount) do
+        with {:ok, store, result} <- do_bump(store, :increment, key, amount) do
+          case result do
+            :ok                  -> {:ok, store, :ok}
+            {:error, no_integer} -> {:ok, store, {:error, :no_integer}}
+          end
+        end
+      end
+      defoverridable bump: 3
 
       @doc false
-      def decrement(store, key, amount \\ 1), do: do_bump(store, :decrement, key, amount)
+      def bump!(store, key, amount) do
+        with {:ok, store, result} <- do_bump(store, :increment, key, amount) do
+          case result do
+            :ok                  -> {:ok, store}
+            {:error, no_integer} -> {:raise, ArithmeticError, [message: msg_for(no_integer, key)]}
+          end
+        end
+      end
+      defoverridable bump!: 3
+
+      defp msg_for(:amount, _key), do: "value provided to operation is not an integer"
+      defp msg_for(:value, key),   do: "value at key #{Inspect.inspect(key, %Inspect.Opts{})} is not an integer"
 
       @doc false
-      def increment!(store, key, amount \\ 1), do: do_bump(store, :increment, key, amount, :raise)
+      def increment(store, key, amount \\ 1) do
+        with {:ok, store, _result} <- do_bump(store, :increment, key, amount) do
+          {:ok, store}
+        end
+      end
+      defoverridable increment: 2, increment: 3
 
       @doc false
-      def decrement!(store, key, amount \\ 1), do: do_bump(store, :decrement, key, amount, :raise)
+      def decrement(store, key, amount \\ 1) do
+        with {:ok, store, _result} <- do_bump(store, :decrement, key, amount) do
+          {:ok, store}
+        end
+      end
+      defoverridable decrement: 2, decrement: 3
 
-      defp do_bump(store, operation, key, amount, raise? \\ false) do
+      defp do_bump(store, operation, key, amount) do
         with {:ok, store, current} <- fetch(store, key) do
           case current do
             :error -> with {:ok, store} <- put(store, key, 0) do
-              do_bump(store, operation, key, amount, raise?)
+              do_bump(store, operation, key, amount)
             end
             {:ok, value} -> case do_bump_calculation(operation, value, amount) do
               {:ok, result} -> with {:ok, store} <- put(store, key, result) do
                 {:ok, store, :ok}
               end
-              {:error, no_integer} -> do_bump_error(store, key, no_integer, raise?)
+              {:error, no_integer} -> {:ok, store, {:error, no_integer}}
             end
           end
         end
