+++ lib/mnemonix/store/map/callbacks.ex
@@ -5,35 +5,6 @@ defmodule Mnemonix.Store.Map.Callbacks do
   defmacro __using__(_) do
     quote location: :keep do
 
-      ####
-      # REQUIRED
-      ##
-
-      def handle_call({:delete, key}, _, store = %__MODULE__{impl: impl}) do
-        case impl.delete(store, key) do
-          {:ok, store}         -> {:reply, :ok, store}
-          {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-        end
-      end
-
-      def handle_call({:fetch, key}, _, store = %__MODULE__{impl: impl}) do
-        case impl.fetch(store, key) do
-          {:ok, store, value}  -> {:reply, {:ok, value}, store}
-          {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-        end
-      end
-
-      def handle_call({:put, key, value}, _, store = %__MODULE__{impl: impl}) do
-        case impl.put(store, key, value) do
-          {:ok, store}         -> {:reply, :ok, store}
-          {:raise, type, args} -> {:reply, {:raise, type, args}, store}
-        end
-      end
-
-      ####
-      # OPTIONAL
-      ##
-
       def handle_call({:fetch!, key}, _, store = %__MODULE__{impl: impl}) do
         case impl.fetch!(store, key) do
           {:ok, store, value}  -> {:reply, {:ok, value}, store}
